Distributed by class. Added a player
